Partial edge edit mode: skip edge nodes reconnection
Add the boolean option 'editEdgeNodes' to the Edge and the
ManipulationSystem, in such a way we can just edit link properties
in editMode without necessarily touching the nodes it connects.

Given that the property is both global at the manipulation level
and local per-edge, we should have enough flexibility for
different use-cases.

The behavior fallbacks to the current one in case nothing is set.