Add example to export to JSON and import it again.
Networks are exported as a JSON array containing an array of objects
representing each node in the network. Each node contains an Id, it's x
and y coordinates and the nodes it is connected to.

Import converts the JSON to a vis DataSet and then regenerates the
network using each node's attributes, inserting it in the same
coordinates and with the same number of connections as before.