Fixed RefException to 'options' object in Items.js
Fixed 2 Reference exceptions to the 'options.editable' object in Items.js.

Given the rest of the code, I believe this was intended to be a check on the editable object of the item's data object. So I have changed it accordingly. It's possible that this is meant to evaluate the options object of the this object, but I don't think so.