Update hierarchy when node level changes
Fix for #3220

On change of data of an existing node, the level is checked for changes.
If changed, for a recalculation of the hierarchical layout.

The fix does not explicitly check for hierarchical layout; this should not be a problem.
The added code can be used to add further node fields which may trigger recalculation of layout.