Update parsing dir to comply with graphviz
This update is for comply with graphviz more strictly. As explained
in [1], graphviz has limitations for combination of arrow[head|tail]
and dir but dotparser.js does not behaviour as same.

[1] https://www.graphviz.org/doc/info/attrs.html#h:undir_note

If attribute list includes 'dir', following cases just be supported.
  1. both or none + arrowhead, arrowtail
  2. forward + arrowhead (arrowtail is not affedted)
  3. back + arrowtail (arrowhead is not affected)

In this update, parsed attributes is pushed 'attr_list' at once and
setValue() is not called in while loop in parseAttributeList().
setValue() is called after while loop ends.
It is because for checking attributes which change the same attribute
of vis. For instance, both of 'arrowhead' and 'dir' (edge style defined
in DOT) make changes for the same 'arrows' attribute in vis. If
setValue() is called from while loop, previous change is overwritten.